增加权限判断

+++ mnmp-domain/src/main/java/com/tarena/mnmp/domain/SmsTemplateDO.java
@@ -55,10 +55,11 @@ public class SmsTemplateDO {
 
     private Integer deleted;
 
-    private Integer createUserId;
+    private Long createUserId;
 
     private String createUserName;
 
     private Long providerId;
+
 
 }
